+++ Makefile
@@ -48,7 +48,9 @@ PARSER_FILES = $(PARSER_DIR)/parser.c \
 			   $(PARSER_DIR)/validator/subshell_validator.c \
 			   $(PARSER_DIR)/utils/free_utils.c \
 			   $(PARSER_DIR)/utils/error_messages.c \
-			   $(PARSER_DIR)/utils/group_utils.c
+			   $(PARSER_DIR)/utils/group_utils.c \
+			   $(PARSER_DIR)/utils/group_sequence_utils.c \
+			   $(PARSER_DIR)/utils/logic_utils.c \
 
 COMMON_FILES = $(COMMON_DIR)/utils.c
